\name{SampleEyeData}
\alias{SampleEyeData}
\docType{data}
\title{
SampleEyeData
}
\description{
The package utilizes a sample of 50 genes from the genes expressions dataset used in Scheetz et al. [2006]. The sampled data set consists of 50 predictors and 40 observations with a continuous response (TRIM32 gene). 
}
\usage{data("SampleEyeData")}
\format{
  A data frame with 40 observations on 51 variables including the response.
}


